requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.7
djangorestframework==3.14.0
Faker==20.1.0
joblib==1.3.2
nltk==3.8.1
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
scikit-learn==1.3.2
scipy==1.11.4
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
textblob==0.17.1
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
wheel==0.41.2
